refactor: HandlerResult와 HttpResponse 통합
