+++ tomcat/src/main/java/org/apache/coyote/http11/handler/Handler.java
@@ -2,6 +2,7 @@ package org.apache.coyote.http11.handler;
 
 import java.io.IOException;
 import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -12,5 +13,5 @@ public abstract class Handler {
     public Handler() {
     }
 
-    public abstract HandlerResult handle(final HttpRequest request) throws IOException;
+    public abstract HttpResponse handle(final HttpRequest request) throws IOException;
 }
