+++ tomcat/src/main/java/org/apache/coyote/http11/response/HttpStatusCode.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.response;
 
 public enum HttpStatusCode {
     OK(200, "OK"),
